+++ src/redux/actions/getLivescoresLatest.js
@@ -4,9 +4,9 @@ import { GET_LIVESCORES_LATEST } from "./actions-type";
 export const getLivescoresLatest = () => {
   return async function (dispatch) {
     try {
-      const response = await axios.get("/livescore/latest");
+      const response = await axios.get("http://localhost:3001/livescore/latest");
       const data = response.data;
-      console.log('la data',data)
+      console.log('la data de livescore desde action',data)
       dispatch({
         type: GET_LIVESCORES_LATEST,
         payload: data,
